refactor: 変数名の修正
